+++ src/main/java/pl/lukaszmatug/transfer/api/resources/ITransferResource.java
@@ -1,5 +1,6 @@
 package pl.lukaszmatug.transfer.api.resources;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 import javax.ws.rs.core.Response;
@@ -7,12 +8,12 @@ import javax.ws.rs.core.Response;
 public interface ITransferResource {
 	public Response getForecast(Optional<String> from, 
 			Optional<String> to,
-			Optional<Double> amount, 
+			Optional<BigDecimal> amount, 
 			Optional<String> feeStrategy);
 	
 	public Response transfer(Optional<String> from, 
 			Optional<String> to,
-			Optional<Double> amount, 
+			Optional<BigDecimal> amount, 
 			Optional<String> feeStrategy);
 
 }
